requirements.txt
python-twitter==3.4.2
configparser==3.5.0
kafka-python==1.4.3